models consolidated
