+++ django_dans_notifications/models/notifications.py
@@ -14,6 +14,24 @@ from django_dans_notifications.logging import LOGGER
 
 """
 # ==================================================================================== #
+# NOTIFICATION BASIC ================================================================= #
+# ==================================================================================== #
+"""
+
+
+#
+# NOTIFICATION BASIC ==================== #
+#
+class NotificationBasic(NotificationBase):
+    read = models.BooleanField(default=False, null=False, blank=False)
+    message = models.CharField(max_length=600, null=False, blank=False)
+
+    def __str__(self):
+        return f"Basic Notification: {self.recipients}"
+
+
+"""
+# ==================================================================================== #
 # NOTIFICATION EMAIL ================================================================= #
 # ==================================================================================== #
 """
@@ -250,3 +268,20 @@ class NotificationEmail(NotificationBase):
 
     def __str__(self):
         return f"Notification Email: {self.sender} -> {self.recipients}"
+
+
+"""
+# ==================================================================================== #
+# NOTIFICATION PUSH ================================================================== #
+# ==================================================================================== #
+"""
+
+
+#
+# NOTIFICATION PUSH ================== #
+#
+class NotificationPush(NotificationBase):
+    message = models.CharField(max_length=300, null=False, blank=False)
+
+    def __str__(self):
+        return f"Notification Push: {self.recipients}"
